server/requirements.txt
pycryptodome
flask
flask_cors
